[verilog] Drop references to TODO and DEPRECATED comment formatting
As pointed out in #6, the section on DEPRECATED comments hsa been
removed from the Google C++ style guide. We currently specifically link
to guidance in that style guide on formatting for TODO and DEPRECATED
comments, but in practice we don't encourage that style - especially not
the TODO(owner) style suggested in the C++ doc.

Fixes #6.