Remove wildcard import restriction
Previously, in the Verilog Style Guide, wildcard import is not
recommended. But the designs at lowRISC repo use the wildcard import in
certain conditions.

This PR is to alleviate the restriction in some conditions.

1. Allow wildcard import at the module header if the imported package is
   in the same IP.
2. Allow wildcard import at the body if the imported package is in the
   same IP.

Please remind that any style of import for the package in the same IP is
not allowed if the import syntax is used out of the module.

```systemverilog {.bad}

import my_pkg::str_t;

module my_module ();
```

This PR closes issue #19

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>