[dv/style] Handling EOT with phase_ready_to_end
This PR updates the section of the DV style guide that relates to
end-of-test practices to now recommend using the phase_ready_to_end()
method of UVM components (the monitor in particular) to control
objections rather than raising and lowering them in the scoreboard.

This also adds some minor modifications to eliminate usage of the word
"you" in several places of the style guide.

Signed-off-by: Udi Jonnalagadda <udij@google.com>