Editorial change to module instantiation section
The module instantiation section was titled "Parameterized Module
Instantiation" and started with guide on instantiating *parametrized*
modules, but then went on to the more general case of instantiating
all kinds of modules. To better match the content of the section, and to
have most general guidance first, rename the section and reshuffle its
contents.